initial commit (with actual functionality, zomg!)
This implements functions for translating a string into "dit"s and
"dah"s on a pin. The end game here is to be able to modulate a CW
transmitter, to serve as a fox-hunting beacon. I may even end up doing
ASK modulation on an analog-out pin, to make an audible tone on an AM
receiver...